+++ wx_explore/common/storage/mongo.py
@@ -116,7 +116,7 @@ class MongoBackend(DataProvider):
 
                     for msg in msgs:
                         # XXX: this only keeps last msg per field breaking ensembles
-                        rows[row_key][f"sf{field_id}"] = zlib.compress(msg[y][x:x+self.n_x_per_row].tobytes())
+                        rows[row_key][f"sf{field_id}"] = zlib.compress(msg[y][x:x+self.n_x_per_row].astype(numpy.float32).tobytes())
 
         with tracing.start_span('put_fields saving') as span:
             self.collection.insert_many(rows.values())
